refactor(Dto): 계층구조 추가
