+++ src/main/java/roomescape/controller/api/AuthController.java
@@ -1,4 +1,9 @@
 package roomescape.controller.api;
 
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
 public class AuthController {
+
 }
