+++ src/main/java/roomescape/controller/api/ThemeController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ThemeCreateRequestDto;
-import roomescape.dto.ThemeResponseDto;
+import roomescape.dto.theme.ThemeCreateRequestDto;
+import roomescape.dto.theme.ThemeResponseDto;
 import roomescape.service.ThemeService;
 
 @RestController
