+++ src/main/java/roomescape/controller/page/AuthPageController.java
@@ -1,4 +1,13 @@
 package roomescape.controller.page;
 
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
 public class AuthPageController {
+
+    @GetMapping("/login")
+    public String loginPage() {
+        return "login";
+    }
 }
