+++ src/main/java/roomescape/dto/auth/LoginRequestDto.java
@@ -1,0 +1,5 @@
+package roomescape.dto.auth;
+
+public record LoginRequestDto(String email, String password) {
+
+}
