+++ src/main/java/roomescape/dto/theme/ThemeCreateRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.theme;
 
 import roomescape.domain.Theme;
